// types/vehicles.ts
export interface Vehicles {
   cargo_capacity: string;
   consumables: string;
   cost_in_credits: string;
   created: Date;
   crew: string;
   edited: Date;
   films: string[];
   length: string;
   manufacturer: string;
   max_atmosphering_speed: string;
   model: string;
   name: string;
   passengers: string;
   pilots: string[];
   url: string;
   vehicle_class: string;
}